# openapi/openapi.yml
openapi: "3.0.3"
info:
  title: "tasks_management"
  version: "1.0"
servers:
  - url: "http://localhost:8080"
components:
  schemas:
    RequestTask:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        completed:
          type: boolean
        deadline:
          type: string
          format: date-time
          example: "2024-09-22T11:57:38+09:00"
    ResponseTask:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        completed:
          type: boolean
        deadline:
          type: string
          format: date-time
          example: "2024-09-22T11:57:38+09:00"
  parameters:
    id:
      name: "id"
      in: "path"
      required: true
      schema:
        type: integer
paths:
  /tasks:
    get:
      tags:
        - "tasks"
      operationId: "getAllTasks"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResponseTask"
    post:
      tags:
        - "tasks"
      operationId: "createTask"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestTask"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseTask"
  /tasks/{id}:
    get:
      tags:
        - "tasks"
      operationId: "getTask"
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseTask"
    put:
      tags:
        - "tasks"
      operationId: "updateTask"
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestTask"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseTask"
    delete:
      tags:
        - "tasks"
      operationId: "deleteTask"
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: "OK。コンテンツなし。"